basic routing and frontend up
