+++ app/controllers/assignments_controller.rb
@@ -1,0 +1,5 @@
+class AssignmentsController < ApplicationController
+  def index
+    skip_policy_scope
+  end
+end
