+++ app/helpers/assignments_helper.rb
@@ -1,0 +1,2 @@
+module AssignmentsHelper
+end
