+++ app/policies/assignment_policy.rb
@@ -1,0 +1,11 @@
+class AssignmentPolicy < ApplicationPolicy
+  class Scope < Scope
+    def resolve
+      scope.all
+    end
+  end
+
+  def index?
+    user.staff?
+  end
+end
