+++ app/views/assignments/index.html.erb
@@ -1,0 +1,2 @@
+
+<%= react_component("ActionItemDashboard", {}) %>
