+++ spec/views/assignments/index.html.erb_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "assignments/index.html.erb", type: :view do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
